Fixed a nasty bug when recompiling a texture definition:
When recompiling a texture definition that is currently being rendered by some
material, there was an error due to the fact that the OpenGL texture buffer was
deleted and a new one was created, however the material's sampler uniforms are
realized only when a material is created.

To solve this, we need to trigger a recompile of materials at the end of the
recompile hook for textures. Therefor we need to track any material names (not
objects because these can change) that are using the texture. We store a list of
material names in the texture objects. We also store a list of texture names in
a material object, and these are kept consistently in sync with each other, so
that whenever we recompile a texture, it recompiles only the materials
referencing it.