Allow colliders to unsubscribe from being picked.
This is useful for full screen colliders used to test if entities are
within the play field. With an OBB the size of the screen and of a
suitable depth to encompass all game objects, it will always be the
closest collider to the picking ray. In this case, the screen OBB
collider could unsubscribe from being picked so that the next closest
collider is picked instead.