Window position is in monitor coordinates, rather than display coordinates.
Display coordinates spans the entire pixel array of all monitors. That
is, with 2 4K monitors side by side, the second monitor's bottom left
pixel is located at 3840x0. When referring to where a window is located,
we want to know that the window is on monitor 2, and at 0x0 within the
space of the monitor's resolution. This changes that.