Fix a potential memory leak caused by keeping a prefab factory's entities hash table populated forever, even after the factory has realized the entity whose entities are valid for.
This could be the case if the slot of some entity in the entities table
stores a reference to some other entity. We could end up with a
circularity of references here, preventing the garbage collector from
ever freeing those entities, even after they are destroyed, and even
after the engine is stopped, since they are accessible from the metadata store.

We just clear the hash table after realizing a prefab to prevent this.
This is fine, because the hash table is repopulated in the factory
lambda that is funcalled when realizing a prefab instance.