Optionally update nodes in parallel.
This is off by default. Users can add `:parallel-updates t` to a
define-config form to enable it.

This is up to the game developer, as it is important GPU state is only
accessed on the main thread. Therefor, any of the game developer's
component update and physics-update hooks need to not have any such GPU
state accesses, and instead defer them until later in a pre-render or
render hook.